Moved all view-related, including sign-in, stuff from MainActivity to HomeFragment.
